ImgOut: option to force individual image indexed PNG export
Previously when we were exporting Indexed PNG files, we would always try to join them into one file provided that they were a total of <256 colors.  This is great for preview sharing, but was annoying for some transformative palette work I was doing with ssf2st stages.  (Consider moving palette set ABC to ordering ACB: that breaks with monolithic indexed PNG storage.)  This new option simply lets you turn off opportunistic joins and instead lets you force individual PNGs per palette.

+++ palmod/PalModDlg_General.cpp
@@ -481,7 +481,7 @@ void CPalModDlg::PostPalSel()
     }
 
     //Reset the prev indexes
-    memset(&nPrevImgIndex[nImgIndexCtr], -1, sizeof(int) * (MAX_IMAGES_DISPLAYABLE - nImgIndexCtr));
+    memset(&nPrevImgIndex[nImgIndexCtr], -1, sizeof(int) * (MAX_IMAGES_DISPLAYABLE - static_cast<size_t>(nImgIndexCtr)));
 
     m_PalHost.EndSetPal();
